Updated waterfallModel description and specs
A description for how to go about modelling in the waterfall style has been added. This defines how we build the model. A section on the specs that will need to be addressed have also been added to define our problem. The next version should include a defined timeline on completing the specs that have been provided.