Updated incrementalModel with outline definition
The prescribed outline stile of incremental modelling has been defined to ensure proper model construction. The specs for the project have been defined so that the phases of the project could be decided on. 4 phases were chosen and headers were created for each. The next commit/work to be done should outline each of the phases for the 4 predicted versions of functionality, as is prescribed by the incremental model.