Updated integrateConfigureModel with definitions and potential modules
The definition for how to create an Integrate and Configure model have been added to ensure the model is constructed right. The CURSE project specs have been added and used to determine the possible modules that the coder might find preconstructed for use. Next commit should include more specifics on the modules that are needed so they can be researched and chosen for construction.