Implemented Step-by-Step Instructions
Given the abstraction of the process in relation to the waterfall model, a strict design has been laid out for the project, documenting the order in which things should happen to achieve a working CURSE program. Next to come is any final changes and tweaks before the design is finalized and potentially implemented.