Restructured Versions and added Version Specs
I changed the description of the class structure from an inherited structure because the users are rather unique. I do not think that that will work since every user type does very different things and wont really inherit anything except for a name and log-ins. I also consolidated the 4th version into the 3rd one because it seemed redundant, both added a finer layer of functionality that was not very different.