Updated resources and completion process
I added some resources that I found helpful for each module. I also added a description on how to bring the project together and build it. Despite the vast resources we have found, I do not feel that this should be the final model because many of the algorithms will have to be customized and work would have to be done at every stage of modules. These resources are still helpful and can be referenced later.