Created UML files
Here I have added the activity models and the use case models for CURSE as well as a use case table for each. These will help with the code that will be constructed at a later date and time, they follow the basic flow of each of the three users while in the program so it outlines the basic functions for the first version modeled.